의존 해결 모델 분리
로그인 완료후 홈화면가짐
