+++ app/src/main/java/com/uuranus/myschedule/LoginActivity.kt
@@ -1,7 +1,6 @@
 package com.uuranus.myschedule
 
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.os.Build
 import android.os.Bundle
 import android.util.Log
