+++ feature/login/src/main/java/com/uuranus/login/LoginScreen.kt
@@ -1,7 +1,6 @@
 package com.uuranus.login
 
 import android.content.Context
-import android.util.Log
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -20,15 +19,12 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.material3.Text
+import androidx.hilt.navigation.compose.hiltViewModel
 import com.uuranus.myschedule.feature.login.R
-import com.kakao.sdk.auth.model.OAuthToken
-import com.kakao.sdk.common.model.ClientError
-import com.kakao.sdk.common.model.ClientErrorCause
-import com.kakao.sdk.user.UserApiClient
 
 
 @Composable
-fun LoginScreen() {
+fun LoginScreen(loginViewModel: LoginViewModel= hiltViewModel()) {
 
     val context = LocalContext.current
 
@@ -63,7 +59,7 @@ fun LoginScreen() {
             Image(
                 painter = painterResource(id = R.drawable.kakao_login_medium_wide),
                 contentDescription = "Login button",
-                modifier = Modifier.fillMaxSize().clickable (onClick = { onClickLogin(context)})
+                modifier = Modifier.fillMaxSize().clickable {loginViewModel.performLogin(context)}
 
 
             )
@@ -71,41 +67,41 @@ fun LoginScreen() {
     }
 }
 
-fun onClickLogin(context: Context) {
-    // 로그인 조합 예제
-
-// 카카오계정으로 로그인 공통 callback 구성
-// 카카오톡으로 로그인 할 수 없어 카카오계정으로 로그인할 경우 사용됨
-    val callback: (OAuthToken?, Throwable?) -> Unit = { token, error ->
-        if (error != null) {
-            Log.e("로그인", "카카오계정으로 로그인 실패", error)
-        } else if (token != null) {
-            Log.i("로그인", "카카오계정으로 로그인 성공 ${token.idToken}")
-            //saveIsLoggedIn(true)
-        }
-    }
-
-// 카카오톡이 설치되어 있으면 카카오톡으로 로그인, 아니면 카카오계정으로 로그인
-    if (UserApiClient.instance.isKakaoTalkLoginAvailable(context)) {
-        UserApiClient.instance.loginWithKakaoTalk(context) { token, error ->
-            if (error != null) {
-                Log.e("로그인", "카카오톡으로 로그인 실패", error)
-
-                // 사용자가 카카오톡 설치 후 디바이스 권한 요청 화면에서 로그인을 취소한 경우,
-                // 의도적인 로그인 취소로 보고 카카오계정으로 로그인 시도 없이 로그인 취소로 처리 (예: 뒤로 가기)
-                if (error is ClientError && error.reason == ClientErrorCause.Cancelled) {
-                    return@loginWithKakaoTalk
-                }
-
-                // 카카오톡에 연결된 카카오계정이 없는 경우, 카카오계정으로 로그인 시도
-                UserApiClient.instance.loginWithKakaoAccount(context, callback = callback)
-            } else if (token != null) {
-                Log.i("로그인", "카카오톡으로 로그인 성공 ${token.idToken}")
-                //saveIsLoggedIn(true)
-            }
-        }
-    } else {
-        UserApiClient.instance.loginWithKakaoAccount(context, callback = callback)
-    }
-}
-
+//fun onClickLogin(context: Context) {
+//    // 로그인 조합 예제
+//
+//// 카카오계정으로 로그인 공통 callback 구성
+//// 카카오톡으로 로그인 할 수 없어 카카오계정으로 로그인할 경우 사용됨
+//    val callback: (OAuthToken?, Throwable?) -> Unit = { token, error ->
+//        if (error != null) {
+//            Log.e("로그인", "카카오계정으로 로그인 실패", error)
+//        } else if (token != null) {
+//            Log.i("로그인", "카카오계정으로 로그인 성공 ${token.idToken}")
+//            //saveIsLoggedIn(true)
+//        }
+//    }
+//
+//// 카카오톡이 설치되어 있으면 카카오톡으로 로그인, 아니면 카카오계정으로 로그인
+//    if (UserApiClient.instance.isKakaoTalkLoginAvailable(context)) {
+//        UserApiClient.instance.loginWithKakaoTalk(context) { token, error ->
+//            if (error != null) {
+//                Log.e("로그인", "카카오톡으로 로그인 실패", error)
+//
+//                // 사용자가 카카오톡 설치 후 디바이스 권한 요청 화면에서 로그인을 취소한 경우,
+//                // 의도적인 로그인 취소로 보고 카카오계정으로 로그인 시도 없이 로그인 취소로 처리 (예: 뒤로 가기)
+//                if (error is ClientError && error.reason == ClientErrorCause.Cancelled) {
+//                    return@loginWithKakaoTalk
+//                }
+//
+//                // 카카오톡에 연결된 카카오계정이 없는 경우, 카카오계정으로 로그인 시도
+//                UserApiClient.instance.loginWithKakaoAccount(context, callback = callback)
+//            } else if (token != null) {
+//                Log.i("로그인", "카카오톡으로 로그인 성공 ${token.idToken}")
+//                //saveIsLoggedIn(true)
+//            }
+//        }
+//    } else {
+//        UserApiClient.instance.loginWithKakaoAccount(context, callback = callback)
+//    }
+//}
+//
